+++ Statistics.h
@@ -2,7 +2,7 @@
 # Classe che implementa i metodi per ottenere #
 # prezzo dell'opzione mediato su tutti gli    #
 # scenari montecarlo (GetPrice) e l'errore    #
-#	montecarlo associato (GetMCError).          #
+# montecarlo associato (GetMCError).          #
 #############################################*/
 
 #ifndef _Statistics_h_
@@ -10,12 +10,12 @@
 
 class Statistics {
 public:
-	__host__ Statistics(float*, float*, int, int);
-	__host__ float GetPrice();
-	__host__ float GetMCError();
+	__host__ Statistics(double*, double*, int, int);
+	__host__ double GetPrice();
+	__host__ double GetMCError();
 private:
-	float _Price;
-	float _MCError;
+	double _Price;
+	double _MCError;
 };
 
 #endif
